requirements.txt
# 提示词优化工具依赖
#python>=3.11
openai==1.82.0
torch~=2.7.0
#triton~=3.3.0
safetensors==0.5.3
torchaudio==2.7.0
sounddevice~=0.5.2
soundfile~=0.13.1
numpy~=2.2.6
funasr~=1.2.6
#realtimetts~=0.5.5
requests~=2.32.3
uvicorn~=0.34.2
fastapi~=0.115.12
starlette~=0.46.2
transformers~=4.52.3
modelscope~=1.26.0
redis==6.2.0
mysql-connector==2.2.9
pyyaml==6.0.2

datasets==3.6.0
tiktoken==0.9.0
wandb==0.19.11
tqdm==4.67.1
chardet==5.2.0

matplotlib==3.10.3
gym==0.26.2
pydub==0.25.1

aiohttp~=3.12.4